+++ core/initializer/test_execution/test_executor/test_executor.go
@@ -9,6 +9,7 @@ import (
 	"context"
 	"fmt"
 	"github.com/kurtosis-tech/container-engine-lib/lib/docker_manager"
+	"github.com/kurtosis-tech/kurtosis-client/golang/kurtosis_core_rpc_api_bindings"
 	"github.com/kurtosis-tech/kurtosis-testsuite-api-lib/golang/kurtosis_testsuite_rpc_api_bindings"
 	"github.com/kurtosis-tech/kurtosis-testsuite-api-lib/golang/kurtosis_testsuite_rpc_api_consts"
 	"github.com/kurtosis-tech/kurtosis/commons/enclave_manager"
@@ -21,6 +22,7 @@ import (
 	"github.com/sirupsen/logrus"
 	"google.golang.org/grpc"
 	"google.golang.org/protobuf/types/known/emptypb"
+	"net"
 	"strings"
 	"time"
 )
@@ -86,7 +88,6 @@ Returns:
 func RunTest(
 		testSetupExecutionCtx context.Context,
 		testsuiteExObjNameProvider *object_name_providers.TestsuiteExecutionObjectNameProvider,
-		initializerContainerId string,
 		log *logrus.Logger,
 		enclaveManager *enclave_manager.EnclaveManager,
 		kurtosisLogLevel logrus.Level,
@@ -129,19 +130,33 @@ func RunTest(
 	dockerManager := enclaveCtx.GetDockerManager()
 	networkId := enclaveCtx.GetNetworkID()
 	kurtosisApiIp := enclaveCtx.GetAPIContainerIPAddr()
-	testsuiteContainerIp := enclaveCtx.GetTestsuiteContainerIPAddr()
+	kurtosisApiHostMachinePortBinding := enclaveCtx.GetAPIContainerHostPortBinding()
 	testsuiteContainerName := enclaveCtx.GetObjectNameProvider().ForTestRunningTestsuiteContainer()
 
-	// We need to a) mount the initializer container inside the enclave network and b) have it register itself with the API container so the
-	//  API container knows that it's a
-	if err := dockerManager.ConnectContainerToNetwork(
+	apiContainerUrlOnHostMachine := fmt.Sprintf(
+		"%v:%v",
+		kurtosisApiHostMachinePortBinding.HostIP,
+		kurtosisApiHostMachinePortBinding.HostPort,
+	)
+	apiContainerConn, err := grpc.Dial(apiContainerUrlOnHostMachine, grpc.WithInsecure())
+	if err != nil {
+		return false, stacktrace.Propagate(err, "An error occurred dialling the API container using its host machine port binding '%v'", apiContainerUrlOnHostMachine)
+	}
+	defer apiContainerConn.Close()
+	apiContainerClient := kurtosis_core_rpc_api_bindings.NewApiContainerServiceClient(apiContainerConn)
+
+	startTestsuiteContainerRegistrationResp, err := apiContainerClient.StartExternalContainerRegistration(
 		testSetupExecutionCtx,
-		networkId,
-		initializerContainerId,
-		ipInsideEnclaveNetwork,
-		"",
-	); err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred connecting container with ID '%v' to the enclave network", containerId)
+		&emptypb.Empty{},
+	)
+	if err != nil {
+		return false, stacktrace.Propagate(err, "An error occurred when starting the registration of the testsuite container")
+	}
+	testsuiteRegistrationKey := startTestsuiteContainerRegistrationResp.RegistrationKey
+	testsuiteIpAddrStr := startTestsuiteContainerRegistrationResp.IpAddr
+	testsuiteIpAddr := net.ParseIP(testsuiteIpAddrStr)
+	if testsuiteIpAddr == nil {
+		return false, stacktrace.NewError("The API container returned an IP address string, '%v', for the testsuite container, but it wasn't parseable to an IP", testsuiteIpAddrStr)
 	}
 
 	/*
@@ -180,23 +195,28 @@ func RunTest(
 		networkId,
 		testsuiteContainerName,
 		kurtosisApiIp,
-		testsuiteContainerIp,
+		testsuiteIpAddr,
 		enclaveId,
 	)
 	if err != nil {
 		return false, stacktrace.Propagate(err, "An error occurred launching the test-running testsuite container")
 	}
-	// No need to have a deferred function that tears down the testsuite container if things don't work because the enclave
-	//  teardown logic that we have earlier will handle it
+	finishRegistrationArgs := &kurtosis_core_rpc_api_bindings.FinishExternalContainerRegistrationArgs{
+		RegistrationKey: testsuiteRegistrationKey,
+		ContainerId:     testsuiteContainerId,
+	}
+	if _, err := apiContainerClient.FinishExternalContainerRegistration(testSetupExecutionCtx, finishRegistrationArgs); err != nil {
+		return false, stacktrace.Propagate(err, "An error occurred finishing the testsuite container registration with the API container")
+	}
 
-	testsuiteEndpointUri := fmt.Sprintf("%v:%v", testsuiteContainerIp.String(), kurtosis_testsuite_rpc_api_consts.ListenPort)
+	testsuiteEndpointUri := fmt.Sprintf("%v:%v", testsuiteIpAddr.String(), kurtosis_testsuite_rpc_api_consts.ListenPort)
 	// TODO SECURITY: Use HTTPS
-	conn, err := grpc.Dial(testsuiteEndpointUri, grpc.WithInsecure())
+	testsuiteContainerConn, err := grpc.Dial(testsuiteEndpointUri, grpc.WithInsecure())
 	if err != nil {
 		return false, stacktrace.Propagate(err, "An error occurred dialing the testsuite container endpoint")
 	}
-	defer conn.Close()
-	testsuiteServiceClient := kurtosis_testsuite_rpc_api_bindings.NewTestSuiteServiceClient(conn)
+	defer testsuiteContainerConn.Close()
+	testsuiteServiceClient := kurtosis_testsuite_rpc_api_bindings.NewTestSuiteServiceClient(testsuiteContainerConn)
 
 	if err := waitUntilTestsuiteContainerIsAvailable(testSetupExecutionCtx, testsuiteServiceClient); err != nil {
 		return false, stacktrace.Propagate(err, "An error occurred while waiting for the testsuite container to become available")