+++ core/docs/changelog.md
@@ -6,6 +6,7 @@
 ### Changes
 * Upgraded to container-engine-lib 0.4.0, which replaces the long list of `CreateAndStartContainer` args with a builder
 * Absorb `kurtosis-core-launcher-lib` into here
+* All execution IDs (sandbox and testing) are now in the format `KTYYYY-MM-DDTHH.MM.SS.sss`
 
 ### Fixes
 * Actually depend on Kurt Client 0.17.1